# api/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "API for managing sandboxes, organizations, billing, and hosts"
  title: Fluid API
  version: "1.0"
servers:
- url: //localhost:8081/v1
paths:
...
  /orgs/{slug}/sources/{vm}/prepare:
    post:
      description: Prepare a source VM for sandbox cloning
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Source VM name
        in: path
        name: vm
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.PrepareRequest"
        description: SSH credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Prepare source VM
      tags:
      - Source VMs
      x-codegen-request-body-name: request
  /orgs/{slug}/sources/{vm}/read:
    post:
      description: Read a file from a source VM
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Source VM name
        in: path
        name: vm
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.ReadSourceRequest"
        description: File path
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orchestrator.SourceFileResult"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Read source file
      tags:
      - Source VMs
      x-codegen-request-body-name: request
  /orgs/{slug}/sources/{vm}/run:
    post:
      description: Execute a read-only command on a source VM
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Source VM name
        in: path
        name: vm
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orchestrator.RunSourceRequest"
        description: Command to run
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orchestrator.SourceCommandResult"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Run source command
      tags:
      - Source VMs
      x-codegen-request-body-name: request
  /orgs/{slug}/vms:
    get:
      description: List all source VMs across connected hosts
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List source VMs
      tags:
      - Source VMs
...
components:
  schemas:
...
    orchestrator.PrepareRequest:
      properties:
        ssh_key_path:
          type: string
        ssh_user:
          type: string
      type: object
    orchestrator.ReadSourceRequest:
      properties:
        path:
          type: string
      type: object
...
    orchestrator.RunSourceRequest:
      properties:
        command:
          type: string
        timeout_seconds:
          type: integer
      type: object
...
    orchestrator.SourceCommandResult:
      example:
        stdout: stdout
        exit_code: 0
        stderr: stderr
        source_vm: source_vm
      properties:
        exit_code:
          type: integer
        source_vm:
          type: string
        stderr:
          type: string
        stdout:
          type: string
      type: object
    orchestrator.SourceFileResult:
      example:
        path: path
        content: content
        source_vm: source_vm
      properties:
        content:
          type: string
        path:
          type: string
        source_vm:
          type: string
      type: object
...
    rest.swaggerError:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
...